+++ packages/cli/src/__tests__/unit/jobrunner.spec.ts
@@ -1,23 +1,26 @@
 import path from "path";
 import { GetPathToTestWrappers } from "@polywrap/test-cases";
-import { ClientConfigBuilder, IClientConfigBuilder } from "@polywrap/client-config-builder-js";
+import {
+  PolywrapClientConfigBuilder,
+  ClientConfigBuilder,
+} from "@polywrap/client-config-builder-js";
 import { testCases } from "./jobrunner-test-cases";
 import { JobRunner } from "../../lib";
 
 jest.setTimeout(200000);
 
 describe("workflow JobRunner", () => {
-  let configBuilder: IClientConfigBuilder;
+  let configBuilder: ClientConfigBuilder;
 
   beforeAll(async () => {
-    configBuilder = new ClientConfigBuilder();
+    configBuilder = new PolywrapClientConfigBuilder();
     const subinvokeUri = `fs/${path.join(
       GetPathToTestWrappers(),
       "subinvoke",
       "00-subinvoke",
       "implementations",
       "rs"
-    )}`
+    )}`;
 
     const invokeUri = `fs/${path.join(
       GetPathToTestWrappers(),
@@ -25,9 +28,11 @@ describe("workflow JobRunner", () => {
       "01-invoke",
       "implementations",
       "rs"
-    )}`
+    )}`;
 
-    configBuilder.addRedirect("ens/imported-invoke.eth", invokeUri).addRedirect("ens/imported-subinvoke.eth", subinvokeUri);
+    configBuilder
+      .setRedirect("ens/imported-invoke.eth", invokeUri)
+      .setRedirect("ens/imported-subinvoke.eth", subinvokeUri);
     configBuilder.addDefaults();
   });