+++ packages/js/manifests/polywrap/src/formats/polywrap/validate.ts
@@ -12,6 +12,8 @@ import {
 import PolywrapManifestSchema_0_1_0 from "@polywrap/polywrap-manifest-schemas/formats/polywrap/0.1.0.json";
 import PolywrapManifestSchema_0_2_0 from "@polywrap/polywrap-manifest-schemas/formats/polywrap/0.2.0.json";
 import PolywrapManifestSchema_0_3_0 from "@polywrap/polywrap-manifest-schemas/formats/polywrap/0.3.0.json";
+import PolywrapManifestSchema_0_4_0 from "@polywrap/polywrap-manifest-schemas/formats/polywrap/0.4.0.json";
+import PolywrapManifestSchema_0_5_0 from "@polywrap/polywrap-manifest-schemas/formats/polywrap/0.5.0.json";
 
 import {
   Schema,
@@ -30,6 +32,8 @@ const schemas: PolywrapManifestSchemas = {
   "0.1.0": PolywrapManifestSchema_0_1_0,
   "0.2.0": PolywrapManifestSchema_0_2_0,
   "0.3.0": PolywrapManifestSchema_0_3_0,
+  "0.4.0": PolywrapManifestSchema_0_4_0,
+  "0.5.0": PolywrapManifestSchema_0_5_0,
 };
 
 const validator = new Validator();