+++ packages/schema/bind/src/__tests__/test-cases.spec.ts
@@ -2,7 +2,9 @@ import { fetchTestCases } from "./index";
 import {
   bindSchema,
   BindLanguage,
-  BindOutput
+  BindOutput,
+  bindLanguageToWrapInfoType,
+  BindOptions
 } from "../";
 
 import {
@@ -12,10 +14,12 @@ import {
   OutputEntry
 } from "@polywrap/os-js";
 
-import fs from "fs";
+import fs, {existsSync, mkdirSync} from "fs";
 import path from "path";
 
 import { deepCopy } from "./utils";
+
+jest.setTimeout(60000);
 
 describe("Polywrap Binding Test Suite", () => {
   const cases = fetchTestCases();
@@ -39,10 +43,23 @@ describe("Polywrap Binding Test Suite", () => {
           outputDirAbs: testCase.input.outputDirAbs,
         };
 
-        const output = bindSchema({
+        const bindOptions: BindOptions = {
           ...deepCopy(testCase.input),
+          wrapInfo: {
+            ...deepCopy(testCase.input.wrapInfo),
+            type: bindLanguageToWrapInfoType(language as BindLanguage)
+          },
           bindLanguage: language as BindLanguage,
-        });
+        };
+
+        if (language == "wrap-go") {
+          if (!bindOptions.config) {
+            bindOptions.config = {};
+          }
+          bindOptions.config.goModuleName = "github.com/testorg/testrepo";
+        }
+
+        const output = await bindSchema(bindOptions);
 
         const sort = (array: OutputEntry[]): OutputEntry[] => {
           array.forEach((entry) => {
@@ -75,6 +92,42 @@ describe("Polywrap Binding Test Suite", () => {
           ),
           JSON.stringify(expectedOutput, null, 2),
         );
+
+        const paths: string[] = [];
+
+
+        const outputDirectoryEntry = (root: string, entry: OutputEntry) => {
+          const entryPath = path.join(root, entry.name);
+          paths.push(entryPath);
+
+          switch (entry.type) {
+            case "File": {
+              writeFileSync(entryPath, entry.data);
+              break;
+            }
+            case "Directory": {
+              for (const subEntry of entry.data) {
+                if (!existsSync(entryPath)) {
+                  mkdirSync(entryPath, { recursive: true });
+                }
+                outputDirectoryEntry(entryPath, subEntry);
+              }
+              break;
+            }
+            default: {
+              throw Error(
+                  `outputDirectoryEntry: Unknown entry type. Entry: ${JSON.stringify(
+                      entry
+                  )}`
+              );
+            }
+          }
+        };
+
+        for (const entry of output.output.entries) {
+          outputDirectoryEntry(testResultDir, entry);
+        }
+
 
         expect(output).toMatchObject(expectedOutput);
       }