+++ packages/schema/bind/src/bindings/index.ts
@@ -1,11 +1,10 @@
 import { GenerateBindingFn } from "./types";
 import { BindLanguage } from "../";
-import * as AssemblyScript from "./assemblyscript";
 import * as Rust from "./rust";
-import * as Python from "./python";
-import * as TypeScript from "./typescript";
+import * as WrapBindgen from "./wrap-bindgen";
+import * as Golang from "./golang";
 
-export { AssemblyScript, Rust, TypeScript };
+export { Rust, Golang };
 export * from "./types";
 export * from "./utils";
 
@@ -13,18 +12,34 @@ export function getGenerateBindingFn(
   bindLanguage: BindLanguage
 ): GenerateBindingFn {
   switch (bindLanguage) {
-    case "wasm-as":
-      return AssemblyScript.Wasm.generateBinding;
-    case "wasm-rs":
+    case "wrap-as":
+      return WrapBindgen.getGenerateBindingFn(
+        "https://github.com/polywrap/wrap-abi-bindgen/tree/wrap-0.1/implementations/wrap-assemblyscript"
+      );
+    case "wrap-rs":
       return Rust.Wasm.generateBinding;
+    case "wrap-go":
+      return Golang.Wasm.generateBinding;
     case "plugin-ts":
-      return TypeScript.Plugin.generateBinding;
+      return WrapBindgen.getGenerateBindingFn(
+        "https://github.com/polywrap/wrap-abi-bindgen/tree/wrap-0.1/implementations/plugin-typescript"
+      );
     case "plugin-rs":
-      return Rust.Plugin.generateBinding;
+      return WrapBindgen.getGenerateBindingFn(
+        "https://github.com/polywrap/wrap-abi-bindgen/tree/wrap-0.1/implementations/plugin-rust"
+      );
     case "plugin-py":
-      return Python.Plugin.generateBinding;
+      return WrapBindgen.getGenerateBindingFn(
+        "https://github.com/polywrap/wrap-abi-bindgen/tree/wrap-0.1/implementations/plugin-python"
+      );
+    case "plugin-kt":
+      return WrapBindgen.getGenerateBindingFn(
+        "https://github.com/polywrap/wrap-abi-bindgen/tree/wrap-0.1/implementations/plugin-kotlin"
+      );
     case "app-ts":
-      return TypeScript.App.generateBinding;
+      return WrapBindgen.getGenerateBindingFn(
+        "https://github.com/polywrap/wrap-abi-bindgen/tree/wrap-0.1/implementations/app-typescript"
+      );
     default:
       throw Error(`Error: Language binding unsupported - ${bindLanguage}`);
   }