+++ packages/schema/bind/src/bindings/rust/index.ts
@@ -1,4 +1,3 @@
 export * as Wasm from "./wasm";
-export * as Plugin from "./plugin";
 export * as Functions from "./functions";
 export * as Types from "./types";