+++ packages/schema/bind/src/index.ts
@@ -9,6 +9,6 @@ Mustache.escape = (value) => value;
 export * from "./types";
 export * from "./bindings";
 
-export function bindSchema(options: BindOptions): BindOutput {
-  return getGenerateBindingFn(options.bindLanguage)(options);
+export async function bindSchema(options: BindOptions): Promise<BindOutput> {
+  return await getGenerateBindingFn(options.bindLanguage)(options);
 }